- Added new updated UI @AddMessage
