+++ app/src/main/java/com/infomantri/autosms/send/asynctask/BaseAsyncTask.kt
@@ -1,0 +1,24 @@
+package com.infomantri.autosms.send.asynctask
+
+import android.os.AsyncTask
+
+class BaseAsyncTask(private val sendSMSFromDb: SendSMSFromDb) : AsyncTask<Void, Void, String>() {
+
+    interface SendSMSFromDb {
+
+        fun onStarted()
+
+        fun onCompleted()
+    }
+
+    override fun doInBackground(vararg p0: Void?): String {
+        sendSMSFromDb.onStarted()
+
+        return ""
+    }
+
+    override fun onPostExecute(result: String?) {
+        super.onPostExecute(result)
+        sendSMSFromDb.onCompleted()
+    }
+}
