+++ app/src/main/java/com/infomantri/autosms/send/database/MessageLivaDataDao.kt
@@ -1,0 +1,20 @@
+package com.infomantri.autosms.send.database
+
+import androidx.lifecycle.LiveData
+import androidx.room.*
+
+@Dao
+interface MessageLivaDataDao {
+
+    @Insert(onConflict = OnConflictStrategy.REPLACE)
+    suspend fun insertMessage(msg: Message)
+
+    @Query("SELECT * from message_table")
+    fun getAllMessages(): LiveData<List<Message>>
+
+    @Query("SELECT * from message_table where sent == 1")
+    fun getSentMsgCount(): LiveData<List<Message>>
+
+    @Delete
+    fun delete(message: Message)
+}
