--- app/src/main/java/com/infomantri/autosms/sender/constants/AppConstants.kt
@@ -1,10 +1,0 @@
-package com.infomantri.autosms.sender.constants
-
-object AppConstants {
-
-    const val REMINDER_TIMESTAMP = "reminder_timestamp"
-    const val REMINDER_ID = "reminder_id"
-    const val MOBILE_NO_1 = "mobile_no_1"
-    const val MOBILE_NO_2 = "mobile_no_2"
-
-}
