--- app/src/main/java/com/infomantri/autosms/sender/database/Subscribers.kt
@@ -1,16 +1,0 @@
-package com.infomantri.autosms.sender.database
-
-import androidx.room.ColumnInfo
-import androidx.room.Entity
-import androidx.room.PrimaryKey
-
-@Entity(tableName = "subscribers_table")
-data class Subscribers(
-    @ColumnInfo(name = "subscribers_column")
-    val mobileNo: String,
-
-    val isSelected: Boolean = true,
-
-    @PrimaryKey(autoGenerate = true)
-    val id: Int = 0
-)
