+++ app/src/main/res/layout/activity_add_message.xml
@@ -44,7 +44,7 @@
             android:background="@color/white">
 
         <com.google.android.material.textfield.TextInputEditText
-                android:id="@+id/etEnterFeedback"
+                android:id="@+id/etEnterMessage"
                 android:layout_width="match_parent"
                 android:layout_height="wrap_content"
                 android:gravity="start|top"
