+++ app/src/main/res/menu/menu_main.xml
@@ -2,7 +2,7 @@
 <menu xmlns:android="http://schemas.android.com/apk/res/android"
       xmlns:app="http://schemas.android.com/apk/res-auto"
       xmlns:tools="http://schemas.android.com/tools"
-      tools:context="com.infomantri.autosms.sender.activity.HomeActivity">
+      tools:context="com.infomantri.autosms.send.activity.HomeActivity">
 
     <item android:id="@+id/action_set_alarm"
           android:orderInCategory="@android:integer/config_mediumAnimTime"
